Working with std C90, Unix socket working, other fixes

+++ src/core/utils/data_types_aliases.h
@@ -14,7 +14,7 @@ typedef long int i32;
 typedef long long int i64;
 
 typedef float f32;
-typedef long float f64;
+typedef double f64;
 
 typedef unsigned char boolean;
